==> oldest/02-identify_CBSA_problem_records.ipynb <==
# default_exp identify_cbsa_errors

# Identify records with incorrect CBSA data
> Print state and city of records with incorrect CBSA data

import pandas as pd

def find_anomalies(yr):
    """Find the records that have CBSA-related data that needs correcting."""
    fname = f'/InfoGroup/data/rurality/df_{yr}_uncorrected.csv'
    df = pd.read_csv(fname,dtype=object)
    unknown = df[(df['CBSA Code']!='00000') & (df['CBSA Level']=='0')]
    return unknown[['ABI','FIPS Code','CBSA Code','CBSA Level']].drop_duplicates()

for year in range(1997,2018):
    bad_cases = find_anomalies(year)
    print(f'\n*** {year} ***')
    print(bad_cases)

# The cells below are examples of what needs to be done for every year in 003-Fix_CBSA

# Dulles, VA is part of the Washington,DC metro area, so CBSA Code should be 47900 and CBSA Level 
# should be 2. It is in Loudon County, VA so its FIPS Code should be 51107.

# Easton, MA is part of the Providence metro area, so CBSA Code should be 39300 and CBSA Level 
# should be 2.  It is in Bristol County, MA so its FIPS Code should be 25005.

# 2017
correcter = {
    'EASTON':['25005','39300','2'],
    'DULLES':['51107','47900','2']
}

for i in unknown.index:
    for k,v in correcter.items():
        if unknown.loc[i]['City'].strip() == k:
            unknown.loc[i]['FIPS Code'] = v[0]
            unknown.loc[i]['CBSA Code'] = v[1]
            unknown.loc[i]['CBSA Level'] = v[2]
