==> oldest/04-OMB_counties.ipynb <==
#default_exp OMB_counties

# The OMB urban concept has two categories: Metropolitan Areas and Micropolitan Areas, both created from counties
# according to criteria described here: 
# https://www.govinfo.gov/content/pkg/FR-2010-06-28/pdf/2010-15605.pdf.

# This notebook, after the file improvements made in the first notebook, defines and maps counties that are
# metropolitan, micropolitan, and neither ("rural").

import pandas as pd

year = 2017
data_file = 'data/df_%d.csv' % year
df = pd.read_csv(data_file,dtype=object)

df.dtypes

len(df[df['CBSA Level'].isnull()])

len(df[df['CBSA Code']=='00000'])

len(df[(df['CBSA Code']=='00000') & (df['CBSA Level']=='0')])

# So CBSA Code of 00000 and CBSA Level are not independent criteria of rurality. CBSA Level == 0 will do.

words = {'0':'rural','1':'micropolitan','2':'metropolitan'}
def make_word(code):
    return words[code]

df['CBSA Level desc'] = df['CBSA Level'].apply(make_word)

def is_it_rural(code):
    if code == '0':
        return '1'
    else:
        return '0'

df['rural_OMB'] = df['CBSA Level'].apply(is_it_rural)

df['rural_OMB'].value_counts()

df['rural_OMB'].value_counts(normalize=True) * 100

# Write a new file
outfile = 'data/df_%d_OMB.csv' % year
df.to_csv(outfile,index=None)
