==> oldest/07-identify_rural_tracts.ipynb <==
#default_exp identify_rural_tracts

# Explore rural census tracts in metro counties (CBSAs) and urban areas (UAs).

import pandas as pd
import geopandas as gpd
from shapely.geometry import Point, Polygon

rural_tracts = [line.rstrip("\n") for line in open("data/rural_census_tracts.lis")]
len(rural_tracts)

# Combine the state-level Census shapefiles. VERY compact code.
from pathlib import Path
pathlist = Path("/InfoGroup/rural/map_files/tracts/").glob("*.shp")

gdf = pd.concat([gpd.read_file(shp) for shp in pathlist]).pipe(gpd.GeoDataFrame)

gdf.to_file("/InfoGroup/rural/map_files/tracts/tl_2017_USA_tract.shp",driver='ESRI Shapefile')

# Compute the centroid of the polygon
gdf['centroid'] = gdf['geometry'].apply(lambda x: Point(x.centroid)) # shapely methods

gdf.columns

gdf.head()

# Now find out if the centroid is inside a UA. UAs do not respect political boundaries so a single
# census tract could be in multiple UAs -- although census tracts are much smaller than UAs and 
# UAs are discontinuous in space, so there must be very few tracts that overlap multiple UAs -- 
# and a single census tract could be fractionally rural or urban. 

# If we don't want to be concerned with fractions of census tracts, dividing up the continuous 
# measures by the same fraction, it is reasonable to say that a 'Census-urban' census tract is 
# one whose centroid falls within a particular UA and to call that UA the one to which it belongs.

# We know that all census tracts are wholly within or wholly outside a CBSA because they do not 
# cross county boundaries. So 'OMB-rural' census tracts are in 'OMB-rural' counties.

# "Each census tract contains at least one BG [ed.,block group], and BGs are uniquely numbered 
# within the census tract. Within the standard census geographic hierarchy, BGs never cross state,
# county, or census tract boundaries but may cross the boundaries of any other geographic entity."

# File that contains polygon geometry of UAs for 2017
ua_gdf = gpd.read_file('map_files/tl_2017_us_uac10.shp')
ua_gdf.columns

ua_gdf.head()

print(len(gdf))
print(len(ua_gdf))

%%time
# Count the overlap of tracts in UAs 
overlap_dict = dict.fromkeys(ua_gdf['GEOID10'].tolist(),0)
for i,ua in ua_gdf[['GEOID10','geometry']].iterrows():
    for j,ct in gdf['centroid'].iteritems():
        if ua['geometry'].contains(ct):
                overlap_dict[ua['GEOID10']] += 1
                break

cdf = pd.DataFrame.from_dict(overlap_dict,orient='index',columns=['count'])

# This shows that no UA contains the centroid of more than one census tract.
cdf['count'].value_counts()

# See the python program tracts.py in /InfoGroup/rural/points-in-polygons/ for execution of this code
# to assign a UA code to each census tract whoise centroid falls within a UA.
%%time
# Assign a UA code 
gdf['UA Code'] = '99999'
for j,ct in gdf.iterrows():
    for i,ua in ua_gdf.iterrows():
        if ct['centroid'].within(ua['geometry']):
            gdf.loc[j]['UA Code'] = ua['GEOID10']
            break

# -----------------------

# InfoGroup for 2017 has data on 82,385 census tracts for all. The HRSA/FORHP file 
# of rural units lists 2,302 rural census tracts in addition to all those in non-Metro counties. The Census's 
# Zip Code-to-Census Tracts relationship file identifies 74,091 census tracts in all states(below).

# The ID of a census tract is the combination of the state FIPS, county FIPS, and tract number.
infile = '/home/tflory/Relationship_Files/Census_Tract_to_PUMA.csv'
ct_df = pd.read_csv(infile,usecols=['STATEFP','COUNTYFP','TRACTCE']).drop_duplicates()

len(ct_df['STATEFP'] <= 56)
