==> oldest/08-rural_by_tract_centroid.ipynb <==
#default_exp rural_by_tract_centroid

# Explore rural census tracts in metro counties (CBSAs) and urban areas (UAs).

import pandas as pd
import geopandas as gpd
from shapely.geometry import Point, Polygon

pd.options.display.max_rows = 999

# All tracts
gdf = gpd.read_file('/InfoGroup/rural/map_files/tracts/tl_2017_USA_tract.shp',dtype=object)

gdf.columns

# Compute the centroid of the polygon
gdf['centroid'] = gdf['geometry'].apply(lambda x: Point(x.centroid)) # shapely methods

gdf.head()

# File that contains polygon geometry of UAs for 2017
#ua_gdf = gpd.read_file('map_files/tl_2017_us_uac10.shp')
# File that contains polygon geometry of UAs for 2010
ua_gdf = gpd.read_file('map_files/tl_2010_us_uac10.shp')
ua_gdf.columns

ua_gdf['UATYP10'].value_counts()

print(len(gdf))
print(len(ua_gdf))

## See the comments below.

xgdf = gdf.sample(frac=0.1).copy()
xua_gdf = ua_gdf.sample(frac=0.1).copy()

print(len(xgdf))
print(len(xua_gdf))

# Assign a UA code 
relationship = []
for j,k in xgdf.iterrows():
    for i,l in xua_gdf.iterrows():
        if k['centroid'].within(l['geometry']):
            relationship.append([k['GEOID'],l['GEOID10'],l['UATYP10']])

rel_df = pd.DataFrame(relationship,columns=['tract GEOID','UA GEOID10','UATYP10'])

rel_df.head()

#rel_df.to_csv('/InfoGroup/rural/points-in-polygons/data/rel_df.csv',index=None)

# The above creation of the relationship list of lists and rel_df.csv file worked on a sample from
# the input files.
# The same logic applied to the full UA and tract files was executed in the tracts.py program in
# /InfoGroup/rural/points-in-polygons/. It took 14.67 clock hours to complete. The cells below 
# read the rel_df.csv
# file created by that program execution. There are 50,198 pairs of tract and UA codes. These 
# are cases in which the
# centroid of a tract was found to be contained within the boundaries of a census urban area. 
# Matching back to the
# full list of census tracts, we can identify those as rural whose centroids do not fall within 
# any UA.

rel_df = pd.read_csv('/InfoGroup/rural/points-in-polygons/data/rel_df.csv',dtype=object)

len(rel_df)

## All tracts
gdf = gpd.read_file('/InfoGroup/rural/map_files/tracts/tl_2017_USA_tract.shp',dtype=object)

# Merge the two dataframes. The resulting dataframe will have the rural tracts flagged by a missing value
# for 'UA Code'.
merged = gdf.merge(rel_df,how='left',left_on='GEOID',right_on='tract_GEOID',indicator=True)

merged['_merge'].value_counts()

rur = merged[merged['_merge']=='left_only']
print(len(rur))
print(rur.head())

merged['rural_spatial_tract'] = merged['UA_GEOID10'].astype(str).apply(lambda x: 1 if x=='nan' else 0)

merged['rural_spatial_tract'].value_counts()

merged.columns

merged.drop(columns=['tract_GEOID','_merge'],inplace=True)

len (merged)

rural_only = merged[merged['rural_spatial_tract']==1]
rural_only.to_csv('/InfoGroup/rural/points-in-polygons/data/rural_spatial_tracts.csv',index=None)

merged.to_csv('/InfoGroup/rural/points-in-polygons/data/all_tracts.csv',index=None)

len(rural_only)/len(merged)

# 31.3% of census tracts are rural by this definition; i.e., that the tract's spatial centroid does not fall
# anywhere within the boundaries of a census urban area.
#
# Do the FORHP rural census tracts (defined entirely differently) overlap considerably with this complete
# list of rural census tracts?

rural_tracts = [line.rstrip("\n") for line in open("data/rural_census_tracts.lis")]
len(rural_tracts)

forhp_tracts = pd.DataFrame(rural_tracts,columns=['GEOID'])

overlap_merge = merged.merge(forhp_tracts,on='GEOID',how='outer',indicator=True)
overlap_merge['_merge'].value_counts()

# So 95 of the FORHP 2,302 rural census tracts do not qualify as rural by the centroid-in-UA criterion.
95/2302
# That's 4.1%.

# The UA and tract files used in this procedure are from 2017. The FORHP calculations apply to
# 2010, with revisions to the second-tier codes in 2019.
