==> oldest/09-definitive_rural_tracts.ipynb <==
#default_exp definitive_rural_tracts

# Make this the definitive list of HRSA/FORHP rural census tracts rather than those in notebook 4(1).

import pandas as pd

df = pd.read_csv('data/nonmetrocountiesandcts2016.csv', dtype=object)

df.info()

rural_tracts_df = df[~df['CT'].isnull()]
len(rural_tracts_df)

# How well do these 2534 tracts overlap with the smaller number in the separate pdf document
# that was processed in notebook 4(1)?

tracts = []
with open('/InfoGroup/rural/data/rural_census_tracts.lis', 'r') as fin:
    for tract in fin:
        tracts.append(tract.rstrip())
ser = pd.DataFrame(tracts, columns=(['CT']))
print(ser.head())
print(len(ser))

overlap = rural_tracts_df.merge(ser, on='CT', how='outer', indicator=True)
overlap['_merge'].value_counts()

# So the pdf lists 87 tracts that are not in the 2016 spreadsheet, and the
# spreadsheet lists 319 that are not in the pdf.
# The pdf includes only tracts with RUCA codes 4-10, except for a few anomalous cases
# that I think are all in Alaska.
#
# What are the RUCA scores for the tracts in the 2016 'nonmetro tracts and counties'
# spreadsheet? Keep in mind that these are selected tracts within metro counties. All tracts
# in nonmetro counties are considered rural, apparently without respect to any RUCA codes.
rural_tracts_df['RUCA 2010'].value_counts()

# BELOW IS THE IMPORTANT STUFF.

import pandas as pd

# Examine ALL census tracts in the 'ruca2010revised' spreadsheet to replicate the
# criteria of rurality and compile a complete list of census tracts, not just those
# in metro counties.
#
# The spreadsheet with this data is linked on
# https://www.ers.usda.gov/data-products/rural-urban-commuting-area-codes/
full_df = pd.read_csv('data/ruca2010revised.csv', dtype=object)

full_df = full_df.astype({'Tract': 'int',
                          'Primary RUCA Code 2010': 'int',
                          'Land Area (square miles) 2010': 'float',
                          'Population Density (per square mile) 2010': 'float'})

# The following criteria identify tracts "that are extremely large and where
# use of RUCA codes alone fails to account for distance to services and sparse population. In
# response to these concerns, ORHP has also designated as rural census tracts with RUCA codes 2
# or 3 that are at least 400 square miles in area with a population density of no more than 35
# people."
exceptional_tracts = full_df[((full_df['Primary RUCA Code 2010'] == 2) |
                              (full_df['Primary RUCA Code 2010'] == 3)) &
                             (full_df['Population Density (per square mile) 2010'] <= 35.0) &
                             (full_df['Land Area (square miles) 2010'] >= 400.0)]

print(len(exceptional_tracts))
print(exceptional_tracts['Population Density (per square mile) 2010'].max())
print(exceptional_tracts['Land Area (square miles) 2010'].min())
print(exceptional_tracts['Primary RUCA Code 2010'].value_counts())

# Now all other rural census tracts: "Census tracts with
# RUCA codes 4 through 10 are considered rural for the purposes of Rural Health grants. "
regular_tracts = full_df[full_df['Primary RUCA Code 2010'] >= 4]
print(len(regular_tracts))
print(regular_tracts['Primary RUCA Code 2010'].value_counts())

all_rural_tracts = pd.concat([regular_tracts, exceptional_tracts])
print(len(all_rural_tracts))
print(len(full_df))

# These are the 2010 census tracts that qualify as rural by the HRSA/FORHP definition: 18.8%
# of the total.
len(all_rural_tracts)/(len(full_df)) * 100.0

all_rural_tracts.info()

all_rural_tracts.to_csv('/InfoGroup/rural/points-in-polygons/data/rural_HRSA_updated_tracts.csv',index=None)

hrsa_tracts = set(all_rural_tracts['Tract'].tolist())
type(hrsa_tracts)

# Compare this to the number and percentage of rural tracts as defined by purely spatial
# relationships: a tract is rural if its spatial centroid is not contained within the polygon
# of a census urban area.
# 31.3% of census tracts are rural by the spatial definition. That's 73,056 census tracts of
# which 22,858 are rural. The UA and tract lists are from 2017.
# The HRSA/FORHP RUCA scores use 2010 data.

# Now get the enterprise counts and employment numbers from InfoGroup for both of these
# rurality concepts.

infile = 'data/df_2017_OMB_Census_HRSA_FAR.csv'
df = pd.read_csv(infile, dtype=object)

# 'Full Census Tract' is the 11-digit tract ID
df['Full Census Tract'].fillna('999999999', inplace=True)
df['Full Census Tract'] = df['Full Census Tract'].astype(int)

df['rural_updated_HRSA'] = df['Full Census Tract'].apply(
    lambda x: 1 if x in hrsa_tracts else 0)
df['rural_updated_HRSA'].value_counts()

print('Percentage of 2017 InfoGroup enterprises that are rural according to updated HRSA standard: ', end='')
rural = len(df[df['rural_updated_HRSA'] == 1])
total = len(df)
print(str(rural/total * 100.0))

df['Employee Size (5) - Location'] = df['Employee Size (5) - Location'].astype(float)
total_emp = df['Employee Size (5) - Location'].sum()
print("Total employment:", str(total_emp))

total_updated_HRSA = df[df['rural_updated_HRSA']
                        == 1]['Employee Size (5) - Location'].sum()
pct = (total_updated_HRSA / total_emp) * 100.0
print('Updated HRSA rural percentage employment:', str(pct))

# Now the same for the tract-spatial standard.
df_spatial = pd.read_csv(
    '/InfoGroup/rural/points-in-polygons/data/all_tracts.csv', dtype=object)
df_spatial['rural_tract'].value_counts()

spatial_tracts = df_spatial[df_spatial['rural_tract']
                            == '1']['GEOID'].astype(int).tolist()

% % time
df['rural_spatial_tracts'] = df['Full Census Tract'].apply(
    lambda x: 1 if x in spatial_tracts else 0)
df['rural_spatial_tracts'].value_counts()

# Percentage of InfoGroup enterprises located in a spatially defined rural census tract.
1793068/len(df) * 100.0

total_spatial_tracts = df[df['rural_spatial_tracts']
                          == 1]['Employee Size (5) - Location'].sum()
pct = (total_spatial_tracts / total_emp) * 100.0
print('Spatially defined rural census tracts percentage employment:', str(pct))

df.to_csv('data/df_2017_OMB_Census_HRSA_FAR_tracts.csv', index=None)

# How much overlap is there between the HRSA-revised and the spatial lists of rural census tracts?

import pandas as pd
df = pd.read_csv('data/df_2017_OMB_Census_HRSA_FAR_tracts.csv', dtype=object)

df.columns

revised_set = set(df[df['rural_updated_HRSA'] == '1']['ABI'].tolist())
spatial_set = set(df[df['rural_spatial_tracts'] == '1']['ABI'].tolist())

print(len(revised_set))
print(len(spatial_set))

% % time
l = 0
r_not_in_s = []
for r in revised_set:
    if r not in spatial_set:
        l += 1
        r_not_in_s.append(r)
print(str(l), 'ABIs in revised_HRSA rural enterprises that are not among spatial rural enterprises.')

% % time
n = 0
s_not_in_r = []
for s in spatial_set:
    if s not in revised_set:
        n += 1
        s_not_in_r.append(r)
print(str(n), 'ABIs in spatial rural enterprises that are not among revised_HRSA rural enterprises.')

total_set = set(list(revised_set) + list(spatial_set))

len(total_set)

len(df)

df['Employee Size (5) - Location'] = df['Employee Size (5) - Location'].astype(float)
total_tracts = df[(df['rural_spatial_tracts'] == '1') | (
    df['rural_updated_HRSA'] == '1')]['Employee Size (5) - Location'].sum()
pct = (total_tracts / 159762888.0) * 100.0
print('Total rural census tracts percentage employment:', str(pct))
