==> oldest/11-tracts_from_shapefiles.ipynb <==
#default_exp tracts_from_shapefiles

import pandas as pd
import geopandas as gpd

# Create census tract data file from state shapefiles. Start with a GeoDataFrame
from pathlib import Path
pathlist = Path("/InfoGroup/rural/map_files/tracts/").glob("*.shp")
gdf = pd.concat([gpd.read_file(shp) for shp in pathlist]).pipe(gpd.GeoDataFrame)
print(len(gdf))

# The list of tracts with primary and secondary RUCA codes, pop, land area
tract_df = pd.read_csv('/InfoGroup/rural/data/ruca2010revised.csv',dtype=object)
len(tract_df)

tract_df.columns

# tracts in 2017 that have their centroid inside an urban area in 
rel_df = pd.read_csv('/InfoGroup/rural/points-in-polygons/data/rel_df.csv',dtype=object)

# This is the file with 2010 urban areas.
len(rel_df)

# This is the file with 2017 urban areas.
len(rel_df)

rel_df.columns

rel_df['UATYP10'].value_counts()

# tract_df identifier = 'Tract'
# rel_df identifier = 'tract_GEOID'
merged = tract_df.merge(rel_df,left_on='Tract',right_on='tract_GEOID',how='outer',indicator=True)

len(merged)

merged['_merge'].value_counts()

# All 2010 census tracts
combined = merged[merged['_merge'] != 'right_only']
len(combined)

combined.columns

combined = combined[['Tract','State-County FIPS Code','State','County','Primary RUCA Code 2010',
                    'Secondary RUCA Code 2010','Tract Population 2010',
                     'Land Area (square miles) 2010','Population Density (per square mile) 2010',
                    'UA_GEOID10','UATYP10','NAMELSAD10']]

# Add the name of the urban area. Should have done this in tracts.py and will add that to
# the script so that these next two cells won't have to be executed next time.
ua_gdf = gpd.read_file('map_files/tl_2010_us_uac10.shp')
ua = ua_gdf[['GEOID10','NAMELSAD10']]
final = combined.merge(ua,left_on='UA_GEOID10',right_on='GEOID10',how='left',indicator=True)

#final['_merge'].value_counts()
#len(final)
#final = final[['Tract','State-County FIPS Code','State','County','Primary RUCA Code 2010',
#                    'Secondary RUCA Code 2010','Tract Population 2010',
#                     'Land Area (square miles) 2010','Population Density (per square mile) 2010',
#                    'UA_GEOID10','UATYP10','NAMELSAD10']]
#final.columns
#final.head(50)

#combined.to_csv('/InfoGroup/rural/data/tracts_with_UA_code.csv',index=None)
final.to_csv('/InfoGroup/rural/data/tracts_with_UA_code.csv',index=None)
