==> oldest/12-frontier_and_remote.ipynb <==
#default_exp frontier_and_remote

# Although both the FAR dataset and the ZCTA-to-whatever relationship files both come from Census,
# the zip code areas presented by FAR are from ESRI and are not entirely compatible with the ZCTA areas.
# ESRI business location data come from InfoGroup. It isn't clear where the zip code data that FAR take
# from ESRI are from, but they almost certainly are postal zip codes.

# From Wikipedia:
# ZCTAs are generalized area representations of the United States Postal Service (USPS) ZIP code service areas, 
# but are not the same as ZIP codes. Individual USPS ZIP codes can cross state, place, county, census tract, 
# census block group and census block boundaries, so the Census Bureau asserts that "there is no correlation 
# between ZIP codes and Census Bureau geography".[1] Moreover, the USPS frequently realigns, merges, or splits 
# ZIP codes to meet changing needs. These changes are usually not reflected in the annual TIGER releases. 
# Each ZCTA is constructed by aggregating the Census 2010 blocks whose addresses use a given ZIP code. In 
# assembling census statistical units to create ZCTAs, the Census Bureau took the ZIP code used by the majority 
# of addresses in each census unit at the time the data was compiled. As a result, some addresses end up with a 
# ZCTA code that is different from their ZIP code. ZCTAs are not developed for ZIP codes that comprise only a 
# small number of addresses.[2] Several ZCTAs represent ZIPs that no longer exist due to realignment by the USPS.

# There are approximately 42,000 ZIP Codes and 32,000 ZCTAs. The main reason that there is not one ZCTA for every 
# ZIP Code is that PO Boxes are excluded in ZCTAs, since only populated areas are included in the Census data.

# Summary: problems using zip code as an analytical unit.

# Postal zip codes change to suit the Postal Services needs and those changes are not immediately reflected
# in other data bases or in TIGER files.

# ZCTA (standardized zipcode-based tablulation areas) units are not the same as postal zip codes, as described
# above. 

# We do not know the source of InfoGroup's zip code data. It is probably self-identified by the enterprise and
# therefore probably a postal zip code.

# ZCTA relationship files do not include any relationship between ZCTAs and postal zip codes. 

# ESRI business location data comes from InfoGroup, which probably suggests that all their zip code data are
# postal zip codes. So the zip codes in the FAR data should be compatible with our InfoGroup data, although not 
# aligned in time. The FAR data comes otherwise from the 2010 Census.

# The Census Geocoder can associate ZCTAs with lat/long, which would be just the thing. But it allows only 1,000 
# individual requests at a time. That would mean 14,700 separate batch jobs to apply ZCTA IDs to the 14.7 mn
# InfoGroup records for 2017. That doesn't come close to scaling.

# Python has a free zipcode database (installable with pip) called 'uszipcode'. It refers only to
# postal zip codes. It does not take addresses as input but it does take lat/long, among many other parameters.
# https://uszipcode.readthedocs.io/index.html#id8

# Through the ZCTA relationship files we can associate ZCTAs with the OMB's 
# metropolitan/micropolitan/nonmetropolitan spatial units and draw a basemap of ZCTAs, but we don't have
# analytical data at the moment at that geographical level.

import pandas as pd

infile = 'data/FARcodesZIPdata2010WithAKandHI.csv'
df = pd.read_csv(infile,dtype=object)

df['ZIP'] = df['ZIP'].apply(lambda x: x.zfill(5) if len(x) < 5 == 0 else x)

df.columns

df.head()

print('***',str(len(df)),'***')
f1 = df[df['far1']=='1']
print(len(f1))
f2 = df[df['far2']=='1']
print(len(f2))
f3 = df[df['far3']=='1']
print(len(f3))
f4 = df[df['far4']=='1']
print(len(f4))

# Create a 'FAR Level' variable: 
# 0 if all far* variables are zero.
# 1 if far1==1 and all others 0.
# 2 if far2==1 and far3==0 and far4==0
# 3 if far3==1 and far4==0
# 4 if far4==1

df_zip = df[['ZIP','far1','far2','far3','far4']].copy()
df_zip[['far1','far2','far3','far4']] = df_zip[['far1','far2','far3','far4']].astype(int)

def farlevel(row):
    if sum([row['far1'],row['far2'],row['far3'],row['far4']]) == 0:
        return 0
    elif row['far1'] == 1 and sum([row['far2'],row['far3'],row['far4']]) == 0:
        return 1
    elif row['far2'] == 1 and sum([row['far3'],row['far4']]) == 0:
        return 2
    elif row['far3'] == 1 and row['far4'] == 0:
        return 3
    elif row['far4'] == 1:
        return 4

df_zip['FAR Level'] = df_zip.apply(farlevel,axis=1)

print(len(df_zip))
df_zip = df_zip.drop_duplicates()
print(len(df_zip))

df_zip['FAR Level'].value_counts()

consistent_0 = df[(df['far1']=='0') & (df['far2']=='0') & (df['far3']=='0') & (df['far4']=='0')]

consistent_1 = df[(df['far1']=='1') & (df['far2']=='1') & (df['far3']=='1') & (df['far4']=='1')]

print(len(consistent_0))
print(len(consistent_1))

consistent = pd.concat([consistent_0,consistent_1],ignore_index=True)

len(consistent)

merge_temp = df.merge(consistent,how='left',on='ZIP',indicator=True)

merge_temp['_merge'].value_counts()

inconsistent = merge_temp[merge_temp['_merge']=='left_only'].drop(['state_y','name_y','far1_y', 'far2_y', 'far3_y', 
        'far4_y', 'gridpop_y', 'sqmi_y', 'density_y', 'fr1pop_y', 'fr2pop_y', 'fr3pop_y', 'fr4pop_y', 'fr1pct_y',
        'fr2pct_y', 'fr3pct_y', 'fr4pct_y', '_merge'], axis=1)

inconsistent.rename(index=str,columns={'state_x':'state','name_x':'name','far1_x':'far1','far2_x':'far2',
                                      'far3_x':'far3','far4_x':'far4','gridpop_x':'gridpop','sqmi_x':'sqmi',
                                      'density_x':'density','fr1pop_x':'fr1pop','fr2pop_x':'fr2pop',
                                      'fr3pop_x':'fr3pop','fr4pop_x':'fr4pop','fr1pct_x':'fr1pct',
                                      'fr2pct_x':'fr2pct','fr3pct_x':'fr3pct','fr4pct_x':'fr4pct'},inplace=True)

print(consistent.columns)
print(inconsistent.columns)

# Created a dataframe of inconsistent far codes to try to figure out what the meaning of them is. 
print(inconsistent.head())
print(inconsistent.tail())

# ========================================
# Testing python's 'uszipcode' database.

# uszipcode has two backend database, SimpleZipcode and Zipcode. Zipcode has more info, but the database file 
# is 450MB (takes more time to download). SimpleZipcode doesn’t has all data points listed above, but the 
# database file is smaller (9MB). By default uszipcode uses SimpleZipcode. You can use this code to choose to 
# use the rich info Zipcode:

#>>> from uszipcode import SearchEngine
#>>> search = SearchEngine(simple_zipcode=False)

from uszipcode import SearchEngine
search = SearchEngine(simple_zipcode=True)

result = search.by_zipcode('53589')

type(result)

d = result.to_dict()

d

search.query(lat=d['lat'],lng=d['lng'],radius=1)

infile = 'data/df_2017_OMB_Census_HRSA.csv'
df_ig = pd.read_csv(infile,dtype=object)
len(df_ig)

df_ig['ZipCode'] = df_ig['ZipCode'].apply(lambda x: x.zfill(5) if len(x) < 5 == 0 else x)

%%time
# Add FAR codes to each InfoGroup record by matching on zip codes
df_out = df_ig.merge(df_zip,how='left',left_on='ZipCode',right_on='ZIP',indicator=True)

df_out['_merge'].value_counts()

df_out.drop(columns=['ZIP','_merge'],inplace=True)

# Select records only for 48 continental states for reporting. (FAR codes apply only to them.)
df_select = df_out[(df_out['Continental']=='1') & (df_out['State FIPS'] != '11')].copy()
len(df_select)

f = len(df_select)
print('***',str(f),'***')

f1 = len(df_select[df_select['FAR Level'] == 1])
print(f1/f)

f2 = len(df_select[df_select['FAR Level'] == 2])
print(f2/f)

f3 = len(df_select[df_select['FAR Level'] == 3])
print(f3/f)

f4 = len(df_select[df_select['FAR Level'] == 4])
print(f4/f)

fmiss = len(df_select[df_select['FAR Level'].isnull()])
print(str(fmiss/f))

fnum = len(df_select[df_select['FAR Level'] >= 0])
print(str(fnum/f))

print('\n'+'Stats only for in-scope (48 states) and non-missing.')
f0 = len(df_select[df_select['FAR Level'] == 0])
print('Not Far and remote:',str(f0/fnum))

ft = len(df_select[df_select['FAR Level'] >= 1])
print('Far and remote:',str(ft/fnum))

outfile = 'data/df_2017_OMB_Census_HRSA_FAR.csv'
df_out.to_csv(outfile,index=None)
